planner
checks if previously made turn, if true drives for 0.5s in a straight
line, if false:  checks if it has to make a turn or drive straight. If
it has to make a turn it does. If it has to drive straight it checks if
the distance is greater than 1.5m, if it is it will drive for 1 second,
otherwise it will drive for a time that has to be determined for the
remaining distance.